Userland: Add starter AML parser for ACPI data
Beginning of a parser for the newly exposed DSDT table in
/sys/firmware/acpi. It only implements one opcode, making
it possible to partially parse the DSDT table in my computer/qemu.

Hopefully the parser can be extended into an interpreter and used
by the kernel to get the correct APCI data.

There is lots of debug output at the moment, most of it will be
removed as the parser grows.